+++ src/main/java/com/banbanmoomani/memilog/DAO/MydiaryMapper.java
@@ -1,6 +1,5 @@
 package com.banbanmoomani.memilog.DAO;
 
-import com.banbanmoomani.memilog.DTO.IntegratedDTO;
 import com.banbanmoomani.memilog.DTO.PostDTO;
 import com.banbanmoomani.memilog.DTO.UserDTO;
 import org.apache.ibatis.annotations.Mapper;
